Fix Redis configuration
Move configuration from 'celery' section to 'redis' as Redis is also
used as Django cache backend -- it was not possible to configure that
in settings.ini before.

Ideally Django cache and Celery should be configured separately
as these may use different backends. But in real world we use one
backend anyway -- so let's introduce more settings when we really
need them.