Delete yaml validation [WAL-655]
Change is required mostly by digital ocean.
yaml validation was added because of digital ocean however a format of
user data is not in a safe, which is allowed by a provider. If it is
invalid - a user data script will silently fail in a droplet by default.

OpenStack has no rules on user data format.
Amazon AWS does not handle user data at the moment.